// products/src/interfaces/contractor.ts
export interface IContractor {
  id: string;
  type: string;
  rut: string;
  companyName: string;
  name: string;
  paternalLastName: string;
  maternalLastName: string;
  legalRepresentative: string;
  line: string;
  birthDate: string;
  address: string;
  district: string;
  email: string;
  phone: string;
}